+++ client/components/booking/expandSearch.tsx
@@ -1,7 +1,7 @@
 // TODO: Separate menu for saved locations
 
 import { useState, useContext } from 'react'
-import { UserContext } from '../context/UserContext'
+import { UserContext } from '../../context/UserContext'
 import { useRouter } from 'next/router'
 import {
 	FaHouseUser,
