+++ client/pages/accountSettings.tsx
@@ -5,7 +5,7 @@ import Page from '@/components/ui/page'
 import Section from '@/components/ui/section'
 import AccountInformation from '@/components/account/AccountInformation'
 import { useForm } from 'react-hook-form'
-import { UserContext } from '@/components/context/UserContext'
+import { UserContext } from '@/context/UserContext'
 import { useRouter } from 'next/router'
 import { User } from '@/redux/types'
 import { UpdateUser } from '@/server'
