+++ client/utils/CoordinateToAddress.tsx
@@ -1,0 +1,11 @@
+export async function CoordinateToAddress(coordinates, setLocation) {
+	await fetch(
+		`https://maps.googleapis.com/maps/api/geocode/json?latlng=${coordinates[0]},${coordinates[1]}&key=${process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}`
+	)
+		.then(function (response) {
+			return response.json();
+		})
+		.then(function (data) {
+			setLocation(data.results[0].formatted_address);
+		});
+}
