+++ client/utils/noPoi.tsx
@@ -1,0 +1,14 @@
+
+export function noPoi(visible) {
+	return [
+		{
+			featureType: 'poi',
+			elementType: 'labels',
+			stylers: [
+				{
+					visibility: visible ? 'on' : 'off',
+				},
+			],
+		},
+	];
+}
